+++ packages/chat/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.14.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.14.0-rc.0...@carbon-labs/ai-chat@0.14.0) (2024-09-05)
+
+**Note:** Version bump only for package @carbon-labs/ai-chat
+
+
+
+
+
+# [0.14.0-rc.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.13.0-rc.1...@carbon-labs/ai-chat@0.14.0-rc.0) (2024-09-04)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @carbon/web-components to v2.10.0 ([#223](https://github.com/carbon-design-system/carbon-labs/issues/223)) ([a72c3e4](https://github.com/carbon-design-system/carbon-labs/commit/a72c3e4bc8dbd4063feac9352490ded6819e432c))
+* **deps:** update dependency @carbon/web-components to v2.11.0 ([#227](https://github.com/carbon-design-system/carbon-labs/issues/227)) ([15f1d1f](https://github.com/carbon-design-system/carbon-labs/commit/15f1d1f75f31dbd9cfbb46cbfcb96b02492dbd4d))
+* **deps:** update dependency @carbon/web-components to v2.11.1 ([#230](https://github.com/carbon-design-system/carbon-labs/issues/230)) ([c781c76](https://github.com/carbon-design-system/carbon-labs/commit/c781c76d23d4d268c77ea4dc3f98a38c898f2d51))
+
+
+### Features
+
+* **chat:** complex feedback system ([#249](https://github.com/carbon-design-system/carbon-labs/issues/249)) ([4b795a8](https://github.com/carbon-design-system/carbon-labs/commit/4b795a8062b8da35017ad971a657d6b0eb4952b4))
+
+
+
+
+
 # [0.13.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.13.0-rc.1...@carbon-labs/ai-chat@0.13.0) (2024-07-15)
 
 **Note:** Version bump only for package @carbon-labs/ai-chat